requirements.txt
flask
flask_sqlalchemy
google-generativeai
python-docx
reportlab
python-dotenv
